Added tangible property repair policy
Companies that do not submit financials to the SEC or have their financial statements audited each year could only deduct repair expenses up to $500 under the IRS's de minimis safe harbor.  This safe harbor applies to a large number of small businesses.  The IRS has increased the de minimis safe harbor amount to $2,500 for tax years beginning after 2015.  They also stated that they will not challenge this for earlier years if other conditions are met.  You must have a policy in place at the beginning of the year to take advantage of this safe harbor.